Test added for list HTML view ISSUE #850
